refactor: 优化组件命名
